Address benchmark review cleanup
Restore preset target ownership to isaaclab_tasks, update benchmark changelog wording, rename the RL library descriptor table, clear stateful KPI formatters after finalize, and polish stale benchmark comments.
